+++ CHANGELOG.md
@@ -5,8 +5,7 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
-
-## **WORK IN PROGRESS**
+## 3.2.1-alpha.6 (2025-01-26)
 
 * (theimo1221) Fix some undefined in Command Sources